Created an activity_form_for_me.xml that has the exact same form as my user form one so I can add speed vans and speed camera locations to my firebase cloud.
Added a button for the activity_main.xml also changed the height and created a new linearlayout for the button will also have to look into added the search panel to its own linear layout separate to the mapview so it only closes the search panel and not everything and I also want to keep the hide and show button available at all times.

AndroidManifest.xml I added the form activity_form_for_me.xml to it so I can get redirected there.

Changed somethings in SpeedVanCameraActivity.java to make the form work.

Added firebase dependency to build.gradle