Updated google-services.json file Added firebase implementation Added the basics to upload information to Firebase
Changed the activity_main.xml to have frame layout with a view

MainActivity.java removed the onPostExecute method that I had commented out.
Removed the duration of the route and the formatDuration method
Changed around the visible image button to have it when the user clicks on it when its open it will collapse and the mapview will expand to fill that space left behind by the search panel collapsing
The image button wasn't working first but now it does and its showing up when the search panel is collapsed.
Still having an issue adding firebase to my app some reason its saying its in the wrong region that I think is fixed however it goes to a database but I don't know what one so will have to look into that more